Created a more organised directory structure.
